requirements.txt
cloudscraper>=1.2.71
boto3>=1.28.0
aioboto3>=11.2.0
requests>=2.31.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0
dependency-injector>=4.41.0 